Somnium/requirements.txt
djangorestframework
markdown
django-filter
celery
redis
gunicorn
gevent
pytils
django_extensions
